Log inserts, updates and deletes
